refactor(web): split the CommandBar and Accounts page
CommandBar.tsx  1241 -> 592
  commandBarHelpers.ts   formatters, the synced-field hook, style constants
  CommandBarViews.tsx    the ten per-command view components

Those ten were already separate components — they just shared a file with
the container that switches between them. Moving them is close to pure
relocation; the container keeps all the state.

Accounts.tsx    1079 -> 802
  accountsHelpers.ts     palettes, label maps, hasLoanTerms, the blank form
  AddAccountModal.tsx    the ~210-line add-account form

`emptyNewAccount` became a factory taking the user's privacy default. It
has to stay a factory: as a module constant its `date` would freeze at
whatever day the module was first imported.

Neither page had a test. Both do now, and so does the command bar:

- Accounts opens the add-account modal and asserts its fields are there,
  so a section lost in the move is caught rather than just the title.
- CommandBar opens through the context and types a trade and an expense,
  asserting the Quantity field appears for one and not the other — it
  fails whether the container stops switching or switches to the wrong
  view.

Each was checked by breaking the thing it covers and confirming it fails.

Verified against a clean origin/dev worktree: dev already carries 38
TypeScript errors, and this branch introduces none of its own.
